# lyrics-web-main/public/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "SimpMusic Lyrics API",
        "version": "v1"
    },
    "servers": [
        {
            "url": "https://api-lyrics.simpmusic.org",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/v1/search": {
            "get": {
                "tags": [
                    "Lyrics"
                ],
                "summary": "Search Lyrics",
                "description": "Performs a general search across lyrics content with optional pagination using limit and offset parameters.",
                "operationId": "searchLyrics",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved lyrics matching the search query",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DocsLyricsListResponseSuccess"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No lyrics found for the search query",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DocsErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error occurred while searching lyrics",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DocsErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "LyricResponseDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "videoId": {
                        "type": "string"
                    },
                    "songTitle": {
                        "type": "string"
                    },
                    "artistName": {
                        "type": "string"
                    },
                    "albumName": {
                        "type": "string"
                    },
                    "durationSeconds": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "plainLyric": {
                        "type": "string"
                    },
                    "syncedLyrics": {
                        "type": "string"
                    },
                    "richSyncLyrics": {
                        "type": "string"
                    },
                    "vote": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "contributor": {
                        "type": "string"
                    },
                    "contributorEmail": {
                        "type": "string"
                    }
                },
                "required": [
                    "albumName",
                    "artistName",
                    "contributor",
                    "contributorEmail",
                    "durationSeconds",
                    "id",
                    "plainLyric",
                    "songTitle",
                    "videoId",
                    "vote"
                ]
            },
            "DocsErrorResponse": {
                "type": "object",
                "properties": {
                    "error": {
                        "$ref": "#/components/schemas/ErrorResponseDTO"
                    },
                    "success": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "error",
                    "success"
                ]
            },
            "ErrorResponseDTO": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "boolean"
                    },
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "reason": {
                        "type": "string"
                    }
                },
                "required": [
                    "code",
                    "error",
                    "reason"
                ]
            },
...
            "DocsLyricsListResponseSuccess": {
                "type": "object",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/LyricResponseDTO"
                        }
                    },
                    "success": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "data",
                    "success"
                ]
            }
        }
    }
}